DOMCore/tests/submissions/Ms2ger/Node-hasAttributes.html

<!DOCTYPE html>
<title>Node.hasAttributes</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  var el = document.createElement("foo")
  assert_false(el.hasAttributes())
  el.setAttribute("bar", "baz");
  assert_true(el.hasAttributes())
  el.removeAttribute("bar");
  assert_false(el.hasAttributes())

  assert_false(document.hasAttributes())
  assert_false(document.doctype.hasAttributes())
  assert_false(document.createDocumentFragment().hasAttributes())
  assert_false(document.createTextNode("A").hasAttributes())
  assert_false(document.createComment("A").hasAttributes())
})
</script>
